Update to Hyper 1.1 on hard mode
That is, without taking a dependency on `axum`. This isn't finished; I have this compiler error:

```
error: implementation of `From` is not general enough
  --> crates/twirp/src/test.rs:40:9
   |
40 |         tokio::spawn(test_server_handle_connection(io, router));
   |         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ implementation of `From` is not general enough
   |
   = note: `Box<(dyn StdError + std::marker::Send + Sync + 'static)>` must implement `From<Box<(dyn StdError + std::marker::Send + Sync + '0)>>`, for any lifetime `'0`...
   = note: ...but it actually implements `From<Box<(dyn StdError + std::marker::Send + Sync + 'static)>>`
```

The surprising thing about this error is if you replace the body of the function being called here,
`test_server_handle_connction`, with `todo!()`, without changing the function signature, the type error
goes away.